inet_cksum(): use unsigned types and change return type
Fix Coverity Scan CID 469910 (Overflowed constant)

+++ src/defs.h
@@ -358,7 +358,7 @@ extern char            *inet_name(uint32_t, int);
 extern char            *inet_fmt(uint32_t, char *, size_t);
 extern char            *inet_fmts(uint32_t, uint32_t, char *, size_t);
 extern uint32_t		inet_parse(char *, int);
-extern int		inet_cksum(uint16_t *, uint32_t);
+extern uint16_t		inet_cksum(uint16_t *, uint32_t);
 
 /* prune.c */
 extern struct gtable	*kernel_table;
